refactored code and made styling changes to pw change
